+++ packages/components/choropleth/CHANGELOG.md
@@ -1,3 +1,7 @@
+## `@svizzle/choropleth` (next)
+
+- document `topoToGeo`
+
 ## `@svizzle/choropleth` v0.4.0
 
 - add `projectionId` so that we can pass either a `projection` prop (a projection function) or an id to select a projection from the available ones